requirements.txt
addict
fairscale
future
lmdb
numpy>=1.17
opencv-python
Pillow
pyyaml
requests
scikit-image
scipy
tb-nightly
torch
torchvision
tqdm
yapf
